Recompile for longs. IDs/size now longs. See below.
Recompiled with the new java compiler that switches all Integers for
Longs.
Also did Int->Long for workspace IDs, object IDs, and object size. WSIDs
probably overkill, but it's possible we might have billions of objects
in a ws. Also possible we might want to bump the size limit up to >
2.2G.

All current tests pass. Needs testing for:
compile_typespec
get_module_info

Needs a test to check that passing a long for an object version throws
the appropriate error
Needs to test that the workspace and object ID counters overflow from
int->long nicely, since they're always set via an $inc and not directly.
Will need to directly muck up the counters with the mongo client.

+++ src/us/kbase/workspace/GetModuleInfoParams.java
@@ -41,7 +41,7 @@ public class GetModuleInfoParams {
     @JsonProperty("type")
     private String type;
     @JsonProperty("ver")
-    private Integer ver;
+    private Long ver;
     private Map<String, Object> additionalProperties = new HashMap<String, Object>();
 
     @JsonProperty("mod")
@@ -75,16 +75,16 @@ public class GetModuleInfoParams {
     }
 
     @JsonProperty("ver")
-    public Integer getVer() {
+    public Long getVer() {
         return ver;
     }
 
     @JsonProperty("ver")
-    public void setVer(Integer ver) {
+    public void setVer(Long ver) {
         this.ver = ver;
     }
 
-    public GetModuleInfoParams withVer(Integer ver) {
+    public GetModuleInfoParams withVer(Long ver) {
         this.ver = ver;
         return this;
     }
